% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_poly.R
\name{convert_to_poly}
\alias{convert_to_poly}
\title{Creates Polygons from Coordinates}
\usage{
convert_to_poly(poly_out)
}
\arguments{
\item{poly_out}{Data frame of output from \code{\link{find_real}}}
}
\value{
SpatialPolygonsDataFrame of pools
}
\description{
Function converts data.frame output \code{\link{find_real}} to SpatialPolygons
and saves as SpatialPolygonsDataFrame
}
