% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mid_line.R
\name{create_mid_line}
\alias{create_mid_line}
\title{Creates mid-pool line}
\usage{
create_mid_line(line1, line2)
}
\arguments{
\item{line1}{Spatial Lines object of water surface}

\item{line2}{Spatial Lines object of bed surface}
}
\value{
List of intersection points, lines of water and bed surfaces, mid-channel line and sections of mid-channel line
}
\description{
Function creates mid-pool line from the intersection between vertical lines
and the bed/water surfaces (from \code{\link{find_real}}). This line is the
vertical mean of the intersection points.
Used to create perpendicular lines and Euclidean distance (depth)
}
