% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_perp_line.R
\name{create_perp_line}
\alias{create_perp_line}
\title{Create perpendicular lines}
\usage{
create_perp_line(midLine, b1, b2)
}
\arguments{
\item{midLine}{Output from \code{\link{create_mid_line}}}

\item{b1}{SpatialLines object equivalent to water surface}

\item{b2}{SpatialLines object equivalent to bed surface}
}
\value{
List of full SpatialLines perpendicular lines and the extent that
        intersect with surfaces.
}
\description{
Functions creates lines perpendicular to the mid-pool line (from \code{\link{create_mid_line}}).
[[1]] is full perp line, [[2]] is between the mid-pool line and the bed/water surfaces.
}
