% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary.R
\name{find_boundary}
\alias{find_boundary}
\title{Create raster of boundary}
\usage{
find_boundary(ras, class1, class2, xMin = 64, xMax = 2011, yMin = 275,
  yMax = 651)
}
\arguments{
\item{ras}{Input classified raster}

\item{class1}{Vector of class value above the boundary}

\item{class2}{Vector of class value below the boundary}

\item{xMin}{Minimum x extent of raster (pixel number)}

\item{xMax}{Maximum x extent of raster (pixel number)}

\item{yMin}{Minimum y extent of raster (pixel number)}

\item{yMax}{Maximum y extent of raster (pixel number)}
}
\value{
Raster object containing cells (value = 1) of boundary location
}
\description{
Function returns a raster object identifying the location of the boundary
between two given classes, for a given classified raster.
}
