% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_common.R
\name{find_common}
\alias{find_common}
\title{Find most common factor}
\usage{
find_common(x)
}
\arguments{
\item{x}{Object that can be coerced into vector}
}
\value{
Value of most frequently occurring value
}
\description{
Function returns most frequent value for a given input
}
