% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_depth.R
\name{find_depth}
\alias{find_depth}
\title{Find depth data of pools}
\usage{
find_depth(bsl, wsl, k, xC, yC)
}
\arguments{
\item{bsl}{List of bed surface profiles (output from \code{\link{fit_loess}})}

\item{wsl}{List of water surface profiles (output from \code{\link{fit_loess}})}

\item{k}{Iteration integer}

\item{xC}{Raster of corrected real space x values}

\item{yC}{Raster of corrected real space y values}
}
\value{
Mean, maximum, minimum and standard deviation of pool depths
}
\description{
Function collates the \code{\link{fit_loess}}, \code{\link{find_diff}},
 \code{\link{find_real}}, \code{\link{convert_to_poly}},
\code{\link{create_mid_line}}, \code{\link{create_perp_line}} and \code{\link{find_dist}}
in order to find pool depths for the workingRange images.
}
