% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dist.R
\name{find_dist}
\alias{find_dist}
\title{Find pool depths}
\usage{
find_dist(channel)
}
\arguments{
\item{channel}{SpatialLines object from the output of \code{\link{create_perp_line}}}
}
\value{
List of mean, maximum, minimum and standard deviation of distance
        between water surface and bed surface profiles, equivalent to the pool depth.
}
\description{
Function calculates euclidian distance for all perpendicular lines between
water and bed surface, and the statistics from that.
}
