% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_real.R
\name{find_real}
\alias{find_real}
\title{Find real coordinates of a lines object}
\usage{
find_real(pts, xC, yC)
}
\arguments{
\item{pts}{Matrix or data.frame of line coordinates}

\item{xC}{Raster of corrected real space x values}

\item{yC}{Raster of corrected real space y values}
}
\value{
Data frame of real coordinates
}
\description{
Function corrects the input spatial lines objects, from \code{\link{find_diff}}
into real space coordinates in order to calculate distance.
}
