% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_loes.R
\name{fit_loess}
\alias{fit_loess}
\title{Creates loess object of surface boundaries}
\usage{
fit_loess(pointsListDF)
}
\arguments{
\item{pointsListDF}{Saved output from \code{\link{extract_coord}}; data.frame of points of surface}
}
\value{
List of clearList (list of x and y values used for loessList) and loessList (list of loess objects)
}
\description{
Function takes a data.frame of boundary coordinates and returns list of loess
objects and the corrected coordinates used to fit those lines. ClearList
removes excessively erroneous values, based on stepped gradient.
}
