% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_rasterise.R
\name{image_rasterise}
\alias{image_rasterise}
\title{Convert image to classified raster}
\usage{
image_rasterise(fileName, mod, xmin1 = 64, xmax1 = 2011, ymin1 = 275,
  ymax1 = 601, d = 3)
}
\arguments{
\item{fileName}{Given file name of image}

\item{mod}{randomForests model}

\item{xmin1}{Minimum x extent of raster (pixel number)}

\item{xmax1}{Maximum x extent of raster (pixel number)}

\item{ymin1}{Minimum y extent of raster (pixel number)}

\item{ymax1}{Maximum y extent of raster (pixel number)}

\item{d}{Size of smoothing window}
}
\value{
Saved file of classified raster
}
\description{
Function loads an image, extracts colour bands and smoothes into raster. This
raster is then classified according to a prebuilt randomForests linear model.
This raster is then reclassified into four umbrella classes and saved as a
.csv file with the suffix 'pred.txt'.
}
