% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_txt.R
\name{read_txt}
\alias{read_txt}
\title{Reads image .txts and finds boundary}
\usage{
read_txt(input, class1, class2, folder = "n", xMin = 64, xMax = 2011,
  yMin = 275, yMax = 651)
}
\arguments{
\item{input}{If single file method this is the individual file name}

\item{class1}{Vector of class value above the boundary}

\item{class2}{Vector of class value below the boundary}

\item{folder}{Character string ("y"/"n") as to whole folder or single files respectively}

\item{xMin}{Minimum x extent of raster (pixel number)}

\item{xMax}{Maximum x extent of raster (pixel number)}

\item{yMin}{Minimum y extent of raster (pixel number)}

\item{yMax}{Maximum y extent of raster (pixel number)}
}
\value{
Raster(s) of boundary
}
\description{
Function reads in either all pred.txt files or single files at a time and
runs \code{\link{find_boundary}} to create a raster of boundary values for
the two given classes.
}
