% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_patch.R
\name{remove_patch}
\alias{remove_patch}
\title{Creates raster and removes small clumps}
\usage{
remove_patch(ras, clump_size = 15)
}
\arguments{
\item{ras}{Input classified raster}

\item{n}{Number of pixels clumps to be removed should be smaller than}
}
\value{
Raster with clumps removed
}
\description{
Function takes a classified raster (see \code{\link{image_rasterise}}) and removes
clumps smaller than n pixels from it. Done for the removal of erroneously
identified patches.
}
