% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_txt.R
\name{write_txt}
\alias{write_txt}
\title{Write all .tif files to classified .txt files}
\usage{
write_txt(folderName, vRange, mod)
}
\arguments{
\item{folderName}{Name of folder to run function in}

\item{vRange}{Vector of values from which the working range and ignored photos
can be calculated. Odd values are included, even are excluded.}
}
\value{
List of classified rasters and also save files to folder
}
\description{
Function applies the \code{\link{image_rasterise}} to all .tif files in a given
folder.
}
